IMPALA-13055: Some Iceberg metadata table tests don't assert
Some tests in the Iceberg metadata table suite use the following regex
to verify numbers in the output: [1-9]\d*|0
However, if this format is given, the test unconditionally passes.

This patch changes this format to \d+ and fixes the test results that
incorrectly passed before due to the test not asserting.

Opened IMPALA-13067 to investigate why the test framework works like
this for |0 in the regexes.

Change-Id: Ie47093f25a70253b3e6faca27d466d7cf6999fad
Reviewed-on: http://gerrit.cloudera.org:8080/21394
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>