IMPALA-13131: Azure OpenAI API expects 'api-key' instead of 'Authorization' in the request header
Updated the POST request when communicating with Azure Open AI
endpoint. The header now includes 'api-key: <api-key>' instead of
'Authorization: Bearer <api-key>'.

Also, removed 'model' as a required param for the Azure Open AI api
call. This is mainly because the endpoint contains deployment which
is basically already mapped to a model.

Testing:
- Updated existing unit test as per the Azure API reference
- Manually tested builtin 'ai_generate_text' using an Azure Open AI
deployment.

Change-Id: If9cc07940ce355d511bcf0ee615ff31042d13eb5
Reviewed-on: http://gerrit.cloudera.org:8080/21493
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>